+++ srcs/files/f_size.c
@@ -6,7 +6,7 @@
 /*   By: Thibault Cheneviere <thibault.cheneviere@telecomnancy.eu>            */
 /*                                                                            */
 /*   Created: 2022/11/07 10:31:33 by Thibault Cheneviere                      */
-/*   Updated: 2022/11/07 10:31:35 by Thibault Cheneviere                      */
+/*   Updated: 2022/11/07 19:16:38 by Thibault Cheneviere                      */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 // Verify if files in the path list correspond to the given size
 int verify_files_by_size(const char *path, char *size)
 {
-    int real_size;
+    off_t real_size;
 
     real_size = get_size(size);
 
@@ -46,3 +46,4 @@ int verify_files_by_size(const char *path, char *size)
 
     return 0;
 }
+
